+++ ios/appsflyer_sdk/Sources/appsflyer_sdk/AppsflyerSdkPlugin.swift
@@ -9,7 +9,7 @@ import UIKit
 import Flutter
 
 // Plugin version
-private let kAppsFlyerPluginVersion = "7.0.1"
+private let kAppsFlyerPluginVersion = "7.0.2"
 
 // Flutter channels
 private let afMethodChannel = "af-api"